Change project.yaml to correct syntax, add more testing to unit test for ProjectConfig (#43)
* Add unit test for project config branch list, change project.yaml to correct syntax.

Before: List of branches was not unit-tested (code was correct though).
Now:
- Added unit tests for branch list in project.yaml
- Fixed minor issues
- Changed project.yaml to be actual correct yaml-syntax

Explanation: In YAML, a list of structs works like this:

list:
  - name: element1
    id: 3
    authors:
      - max
      - moritz
  - name: element2
    id: 4
    authors:
      - max
      - moritz

A dash starts a new element with the attribute. Indentation is important!

Optional<T> is not supported by SnakeYAML yet, therefore not supported by Spring-Boot
(since it is using SnakeYAML for parsing). Optional arguments have to be checked
against null.

* Change deployment attributes of branch to "type", "name" and "arguments".

- Arguments are separated by spaces
- webhook-type needs url, script-type needs path to script.